se ha creado el token para login y register, ademas se modifica en dos controladores
